Cleanup reconstruction bindings
